complete

+++ app/Http/Controllers/admin/OrderController.php
@@ -16,10 +16,14 @@ class OrderController extends Controller
         });
     }
 
-    function index()
+    function index(request $request)
     {
-        $orders = Order::latest()->paginate(40);
-        return view('admin.order.index', compact('orders'));
+        if ($request->user()->cannot('view', Order::class)) {
+            return redirect()->back()->with('notification', 'Bạn không có quyền truy cập');
+        } else {
+            $orders = Order::latest()->paginate(40);
+            return view('admin.order.index', compact('orders'));
+        }
     }
 
     function getDetail(request $request)
@@ -42,27 +46,32 @@ class OrderController extends Controller
         return response()->json(['msg' => 'Đã cập nhật trạng thái đơn hàng', 'view' => $view]);
     }
 
-    function delete(request $request) {
-        $order = Order::find($request->id);
-        $order->products()->detach();
-        Order::destroy($request->id);
-        $orders = Order::latest()->paginate(40);
-        $view = view('admin.order.inc.main_data', compact('orders'))->render();
-        return response()->json(['msg' => 'Đã xóa đơn hàng', 'view' => $view]);
+    function delete(request $request)
+    {
+        if ($request->user()->cannot('delete', Order::class)) {
+            return response()->json(['code'=> -1, 'msg'=>'Bạn không có quyền xóa']);
+        } else {
+            $order = Order::find($request->id);
+            $order->products()->detach();
+            Order::destroy($request->id);
+            $orders = Order::latest()->paginate(40);
+            $view = view('admin.order.inc.main_data', compact('orders'))->render();
+            return response()->json(['msg' => 'Đã xóa đơn hàng', 'view' => $view]);
+        }
     }
 
     public function filter(request $request)
     {
         $str = $request->search_string;
-       
+
         $stt = $request->status;
         $orders = Order::where(function ($q) use ($str) {
-                            $q->where("code", "like", "%$str%")
-                            ->orWhere("name", "like", "%$str%")
-                            ->orWhere('created_at', 'like', "$str");
-                        })->when($stt, function($q) use ($stt) {
-                            $q->where('status', $stt);
-                        })->latest()->paginate(40);
+            $q->where("code", "like", "%$str%")
+                ->orWhere("name", "like", "%$str%")
+                ->orWhere('created_at', 'like', "$str");
+        })->when($stt, function ($q) use ($stt) {
+            $q->where('status', $stt);
+        })->latest()->paginate(40);
 
         $view = view('admin.order.inc.main_data', compact('orders'))->render();
         return response()->json(['view' => $view]);
